requirements.txt
selenium
pytest
python-dotenv
allure-pytest